% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{R85P3S03}
\alias{R85P3S03}
\title{Table R85P3S03 \cr
Household clothing expenditure of rural household in 1385 \cr
هزينه هاي پوشاك خانوار روستایی سال 1385}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
R85P3S03
}
\description{
{R for Rural} \cr
{85 for 1385} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
